backend/requirements.txt
flask
pyyaml
flask-cors
ruamel.yaml
apscheduler
google-generativeai
